gazelle: only rename BUILD files in "fix" mode (#343)
Fixes #335

+++ go/tools/gazelle/gazelle/main.go
@@ -108,7 +108,7 @@ func run(dirs []string, emit func(*bzl.File) error, external rules.ExternalResol
 			if err := emit(f); err != nil {
 				return err
 			}
-			if f.Path != existingFilePath {
+			if *mode == "fix" && f.Path != existingFilePath {
 				if err := os.Remove(existingFilePath); err != nil {
 					return err
 				}
